[SR] Rename session-with -> with-req-action-session
